--- a/babel.config.js
+++ b/babel.config.js
@@ -1,5 +1,7 @@
-// the file is used for jest testing & site building
+// the file is used for site building and other tooling builds
 // > 2%, make template string not compiled to concat, since it's not fast
+const process = require('node:process')
+
 module.exports = {
   presets:
     process.env.NODE_ENV === 'test'