--- a/build/utils/terse-cssr.ts
+++ b/build/utils/terse-cssr.ts
@@ -1,8 +1,10 @@
 import { generate } from '@babel/generator'
 import { parse } from '@babel/parser'
-import traverse from '@babel/traverse'
+import _traverse from '@babel/traverse'
 
 export function terseCssr(code: string): string {
+  // https://github.com/babel/babel/discussions/13093
+  const traverse = (_traverse as any)?.default ?? _traverse
   const patternSpace = / +/g
   const patternEnter = /\n+/g