--- a/src/alert/tests/Alert.spec.ts
+++ b/src/alert/tests/Alert.spec.ts
@@ -48,9 +48,10 @@ describe('n-alert', () => {
 
   it('should work with `bordered` prop', async () => {
     const wrapper = mount(NAlert)
-    expect(wrapper.find('.n-alert-body--bordered').exists()).toBe(true)
+    const body = wrapper.find('.n-alert-body')
+    expect(body.classes()).toContain('n-alert-body--bordered')
     await wrapper.setProps({ bordered: false })
-    expect(wrapper.find('.n-alert-body--bordered').exists()).toBe(false)
+    expect(body.classes()).not.toContain('n-alert-body--bordered')
     wrapper.unmount()
   })
 
@@ -127,8 +128,8 @@ describe('n-alert', () => {
   })
 
   it('should trigger callback when closed', async () => {
-    const handleCloseClick = jest.fn()
-    const handleOnAfterLeave = jest.fn()
+    const handleCloseClick = vi.fn()
+    const handleOnAfterLeave = vi.fn()
     const wrapper = mount(NAlert, {
       props: {
         closable: true,
@@ -144,9 +145,9 @@ describe('n-alert', () => {
 
     expect(handleCloseClick).toHaveBeenCalled()
 
-    setTimeout(() => {
+    vi.waitFor(() => {
       expect(handleOnAfterLeave).toHaveBeenCalled()
-      wrapper.unmount()
-    }, 0)
+    })
+    wrapper.unmount()
   })
 })