--- a/src/cascader/tests/Cascader.spec.ts
+++ b/src/cascader/tests/Cascader.spec.ts
@@ -100,14 +100,14 @@ describe('n-cascader', () => {
       ] as const
     ).forEach((placement) => {
       const wrapper = mount(NCascader, { props: { placement } })
-      setTimeout(() => {
+      vi.waitFor(() => {
         expect(
           document
             .querySelector('.v-binder-follower-content')
             ?.getAttribute('v-placement')
         ).toBe(placement)
-        wrapper.unmount()
       })
+      wrapper.unmount()
     })
   })
 
@@ -190,7 +190,7 @@ describe('n-cascader', () => {
   })
 
   it('should work with `on-blur` prop', async () => {
-    const onBlur = jest.fn()
+    const onBlur = vi.fn()
     const wrapper = mount(NCascader, {
       props: { options: getOptions(), onBlur }
     })
@@ -200,7 +200,7 @@ describe('n-cascader', () => {
   })
 
   it('should work with `on-focus` prop', async () => {
-    const onFocus = jest.fn()
+    const onFocus = vi.fn()
     const wrapper = mount(NCascader, {
       props: { options: getOptions(), onFocus }
     })
@@ -217,12 +217,15 @@ describe('n-cascader', () => {
       }
     })
 
-    await wrapper.find('.n-base-selection').trigger('click')
-    expect(wrapper.find('.n-base-selection--active').exists()).toBe(true)
+    const selection = wrapper.find('.n-base-selection')
+
+    await selection.trigger('click')
+    expect(selection.classes()).toContain('n-base-selection--active')
+
     expect(document.querySelector('.n-cascader-menu')).not.toEqual(null)
 
-    await wrapper.find('.n-base-selection').trigger('click')
-    expect(wrapper.find('.n-base-selection--active').exists()).toBe(false)
+    await selection.trigger('click')
+    expect(selection.classes()).not.toContain('n-base-selection--active')
     expect(document.querySelector('.n-cascader-menu')).toEqual(null)
     wrapper.unmount()
   })