--- a/src/code/tests/Code.spec.tsx
+++ b/src/code/tests/Code.spec.tsx
@@ -8,7 +8,7 @@ hljs.registerLanguage('javascript', javascript)
 
 describe('n-code', () => {
   it('should warn when no hljs is set', () => {
-    const spy = jest.spyOn(console, 'error').mockImplementation()
+    const spy = vi.spyOn(console, 'error').mockImplementation(() => {})
     mount(NCode)
     expect(spy).toHaveBeenCalled()
   })