--- a/src/collapse/tests/Collapse.spec.tsx
+++ b/src/collapse/tests/Collapse.spec.tsx
@@ -152,7 +152,7 @@ describe('n-collapse', () => {
   })
 
   it('should work with `on-item-header-click` prop', async () => {
-    const onClick = jest.fn()
+    const onClick = vi.fn()
     const wrapper = mount(NCollapse, {
       props: {
         onItemHeaderClick: onClick