--- a/src/color-picker/tests/ColorPicker.spec.tsx
+++ b/src/color-picker/tests/ColorPicker.spec.tsx
@@ -128,7 +128,7 @@ describe('n-color-picker', () => {
       wrapper.unmount()
     })
     it('output according to mode', async () => {
-      const onUpdateValue = jest.fn()
+      const onUpdateValue = vi.fn()
       const output = {
         RGBA: {
           mode: 'rgb',
@@ -215,14 +215,14 @@ describe('n-color-picker 2', () => {
       ] as const
     ).forEach((placement) => {
       const wrapper = mount(NColorPicker, { props: { placement } })
-      setTimeout(() => {
+      vi.waitFor(() => {
         expect(
           document
             .querySelector('.v-binder-follower-content')
             ?.getAttribute('v-placement')
         ).toBe(placement)
-        wrapper.unmount()
       })
+      wrapper.unmount()
     })
   })