--- a/src/dialog/tests/Dialog.spec.tsx
+++ b/src/dialog/tests/Dialog.spec.tsx
@@ -121,7 +121,7 @@ describe('n-dialog', () => {
   })
 
   it('onMaskClick', async () => {
-    const onMaskClick = jest.fn()
+    const onMaskClick = vi.fn()
     const Test = defineComponent({
       setup() {
         const dialog = useDialog()