--- a/src/divider/tests/server.spec.tsx
+++ b/src/divider/tests/server.spec.tsx
@@ -1,7 +1,7 @@
 import { setup } from '@css-render/vue3-ssr'
 import { renderToString } from '@vue/server-renderer'
 /**
- * @jest-environment node
+ * @vitest-environment node
  */
 import { createSSRApp, h } from 'vue'
 import { NDivider } from '../..'