--- a/src/jest-setup.ts
+++ b/src/jest-setup.ts
@@ -1,6 +1,0 @@
-import { isBrowser } from './_utils'
-
-// https://github.com/jsdom/jsdom/issues/1422
-if (isBrowser) {
-  HTMLDivElement.prototype.scrollTo = () => {}
-}