# Migrate to vitest earning

- 5x dev performance
- 2x CI performance


## local

### Hardware

CPU: Apple M4 Pro (12) @ 4.51 GHz
GPU: Apple M4 Pro (16) @ 1.58 GHz [Integrated]
Memory: 19.21 GiB / 24.00 GiB (80%)

### before

Test Suites: 189 passed, 189 total
Tests:       1007 passed, 1007 total
Snapshots:   199 passed, 199 total
Time:        267.155 s
Ran all test suites.

### after

Test Files  189 passed (189)
Tests  1007 passed (1007)
Start at  21:52:02
Duration  51.12s (transform 10.32s, setup 1.32s, collect 442.04s, tests 33.15s, environment 38.34s, prepare 8.67s)

## CI
### before
[11m 49s](https://github.com/tusen-ai/naive-ui/actions/runs/17999329156/usage)

### after

[5m 9s](https://github.com/tusen-ai/naive-ui/actions/runs/18040492299/usage)